+++ subscriber/podaac_data_subscriber.py
@@ -15,19 +15,17 @@
 import argparse
 import logging
 import os
+import sys
+from datetime import datetime, timedelta
 from os import makedirs
 from os.path import isdir, basename, join, isfile, exists
+from urllib.error import HTTPError
 from urllib.request import urlretrieve
-from datetime import datetime, timedelta
 import hashlib
 
 from subscriber import podaac_access as pa
 
 __version__ = pa.__version__
-
-LOGLEVEL = os.environ.get('PODAAC_LOGLEVEL', 'WARNING').upper()
-logging.basicConfig(level=LOGLEVEL)
-logging.debug("Log level set to " + LOGLEVEL)
 
 page_size = 2000
 
@@ -40,7 +38,9 @@ def get_update_file(data_dir, collection_name):
     if isfile(data_dir + "/.update__" + collection_name):
         return data_dir + "/.update__" + collection_name
     elif isfile(data_dir + "/.update"):
-        print("WARNING: found a deprecated use of '.update' file at {0}. After this run it will be renamed to {1}".format(data_dir + "/.update", data_dir + "/.update__" + collection_name))
+        logging.warning(
+            "found a deprecated use of '.update' file at {0}. After this run it will be renamed to {1}".format(
+                data_dir + "/.update", data_dir + "/.update__" + collection_name))
         return data_dir + "/.update"
 
     return None
@@ -48,7 +48,8 @@ def get_update_file(data_dir, collection_name):
 
 def validate(args):
     if args.minutes is None and args.startDate is False and args.endDate is False:
-        raise ValueError("Error parsing command line arguments: one of --start-date, --end-date or --minutes are required")
+        raise ValueError(
+            "Error parsing command line arguments: one of --start-date, --end-date or --minutes are required")
 
 
 def create_parser():
@@ -56,33 +57,54 @@ def create_parser():
     parser = argparse.ArgumentParser(prog='PO.DAAC data subscriber')
 
     # Adding Required arguments
-    parser.add_argument("-c", "--collection-shortname", dest="collection",required=True, help = "The collection shortname for which you want to retrieve data.")  # noqa E501
-    parser.add_argument("-d", "--data-dir", dest="outputDirectory", required=True, help = "The directory where data products will be downloaded.")  # noqa E501
+    parser.add_argument("-c", "--collection-shortname", dest="collection", required=True,
+                        help="The collection shortname for which you want to retrieve data.")  # noqa E501
+    parser.add_argument("-d", "--data-dir", dest="outputDirectory", required=True,
+                        help="The directory where data products will be downloaded.")  # noqa E501
 
     # Adding optional arguments
     parser.add_argument("-f", "--force", dest="force", action="store_true", help = "Flag to force downloading files that are listed in CMR query, even if the file exists and checksum matches")  # noqa E501
 
     # spatiotemporal arguments
-    parser.add_argument("-sd", "--start-date", dest="startDate", help = "The ISO date time before which data should be retrieved. For Example, --start-date 2021-01-14T00:00:00Z", default=False)  # noqa E501
-    parser.add_argument("-ed", "--end-date", dest="endDate", help = "The ISO date time after which data should be retrieved. For Example, --end-date 2021-01-14T00:00:00Z", default=False)   # noqa E501
-    parser.add_argument("-b", "--bounds", dest="bbox", help = "The bounding rectangle to filter result in. Format is W Longitude,S Latitude,E Longitude,N Latitude without spaces. Due to an issue with parsing arguments, to use this command, please use the -b=\"-180,-90,180,90\" syntax when calling from the command line. Default: \"-180,-90,180,90\".", default="-180,-90,180,90")  # noqa E501
+    parser.add_argument("-sd", "--start-date", dest="startDate",
+                        help="The ISO date time before which data should be retrieved. For Example, --start-date 2021-01-14T00:00:00Z",
+                        default=False)  # noqa E501
+    parser.add_argument("-ed", "--end-date", dest="endDate",
+                        help="The ISO date time after which data should be retrieved. For Example, --end-date 2021-01-14T00:00:00Z",
+                        default=False)  # noqa E501
+    parser.add_argument("-b", "--bounds", dest="bbox",
+                        help="The bounding rectangle to filter result in. Format is W Longitude,S Latitude,E Longitude,N Latitude without spaces. Due to an issue with parsing arguments, to use this command, please use the -b=\"-180,-90,180,90\" syntax when calling from the command line. Default: \"-180,-90,180,90\".",
+                        default="-180,-90,180,90")  # noqa E501
 
     # Arguments for how data are stored locally - much processing is based on
     # the underlying directory structure (e.g. year/Day-of-year)
-    parser.add_argument("-dc", dest="cycle", action="store_true", help = "Flag to use cycle number for directory where data products will be downloaded.")  # noqa E501
-    parser.add_argument("-dydoy", dest="dydoy", action="store_true", help = "Flag to use start time (Year/DOY) of downloaded data for directory where data products will be downloaded.")  # noqa E501
-    parser.add_argument("-dymd", dest="dymd", action="store_true", help = "Flag to use start time (Year/Month/Day) of downloaded data for directory where data products will be downloaded.")  # noqa E501
-    parser.add_argument("-dy", dest="dy", action="store_true", help = "Flag to use start time (Year) of downloaded data for directory where data products will be downloaded.")  # noqa E501
-    parser.add_argument("--offset", dest="offset", help = "Flag used to shift timestamp. Units are in hours, e.g. 10 or -10.")  # noqa E501
+    parser.add_argument("-dc", dest="cycle", action="store_true",
+                        help="Flag to use cycle number for directory where data products will be downloaded.")  # noqa E501
+    parser.add_argument("-dydoy", dest="dydoy", action="store_true",
+                        help="Flag to use start time (Year/DOY) of downloaded data for directory where data products will be downloaded.")  # noqa E501
+    parser.add_argument("-dymd", dest="dymd", action="store_true",
+                        help="Flag to use start time (Year/Month/Day) of downloaded data for directory where data products will be downloaded.")  # noqa E501
+    parser.add_argument("-dy", dest="dy", action="store_true",
+                        help="Flag to use start time (Year) of downloaded data for directory where data products will be downloaded.")  # noqa E501
+    parser.add_argument("--offset", dest="offset",
+                        help="Flag used to shift timestamp. Units are in hours, e.g. 10 or -10.")  # noqa E501
 
-    parser.add_argument("-m", "--minutes", dest="minutes", help = "How far back in time, in minutes, should the script look for data. If running this script as a cron, this value should be equal to or greater than how often your cron runs.", type=int, default=None)  # noqa E501
-    parser.add_argument("-e", "--extensions", dest="extensions", help = "The extensions of products to download. Default is [.nc, .h5, .zip]", default=None, action='append')  # noqa E501
-    parser.add_argument("--process", dest="process_cmd", help="Processing command to run on each downloaded file (e.g., compression). Can be specified multiple times.", action='append')
+    parser.add_argument("-m", "--minutes", dest="minutes",
+                        help="How far back in time, in minutes, should the script look for data. If running this script as a cron, this value should be equal to or greater than how often your cron runs.",
+                        type=int, default=None)  # noqa E501
+    parser.add_argument("-e", "--extensions", dest="extensions",
+                        help="The extensions of products to download. Default is [.nc, .h5, .zip]", default=None,
+                        action='append')  # noqa E501
+    parser.add_argument("--process", dest="process_cmd",
+                        help="Processing command to run on each downloaded file (e.g., compression). Can be specified multiple times.",
+                        action='append')
 
-    parser.add_argument("--version", action="version", version='%(prog)s ' + __version__, help="Display script version information and exit.")  # noqa E501
-    parser.add_argument("--verbose", dest="verbose", action="store_true", help="Verbose mode.")    # noqa E501
+    parser.add_argument("--version", action="version", version='%(prog)s ' + __version__,
+                        help="Display script version information and exit.")  # noqa E501
+    parser.add_argument("--verbose", dest="verbose", action="store_true", help="Verbose mode.")  # noqa E501
 
-    parser.add_argument("-p", "--provider", dest="provider", default='POCLOUD', help="Specify a provider for collection search. Default is POCLOUD.")    # noqa E501
+    parser.add_argument("-p", "--provider", dest="provider", default='POCLOUD',
+                        help="Specify a provider for collection search. Default is POCLOUD.")  # noqa E501
     return parser
 
 
@@ -174,8 +196,8 @@ def run():
         pa.validate(args)
         validate(args)
     except ValueError as v:
-        print(v)
-        exit()
+        logging.error(str(v))
+        exit(1)
 
     pa.setup_earthdata_login_auth(edl)
     token = pa.get_token(token_url, 'podaac-subscriber', edl)
@@ -218,7 +240,7 @@ def run():
     # This cell will replace the timestamp above with the one read from the `.update` file in the data directory, if it exists.
 
     if not isdir(data_path):
-        print("NOTE: Making new data directory at " + data_path + "(This is the first run.)")
+        logging.info("NOTE: Making new data directory at " + data_path + "(This is the first run.)")
         makedirs(data_path, exist_ok=True)
 
     else:
@@ -227,11 +249,12 @@ def run():
             try:
                 with open(update_file, "r") as f:
                     data_within_last_timestamp = f.read().strip()
-                    print("NOTE: Update found in the data directory. (The last run was at " + data_within_last_timestamp + ".)")
+                    logging.info(
+                        "NOTE: Update found in the data directory. (The last run was at " + data_within_last_timestamp + ".)")
             except FileNotFoundError:
-                print("WARN: No .update in the data directory. (Is this the first run?)")
+                logging.warning("No .update in the data directory. (Is this the first run?)")
         else:
-            print("WARN: No .update__" + short_name + " in the data directory. (Is this the first run?)")
+            logging.warning("No .update__" + short_name + " in the data directory. (Is this the first run?)")
 
     # Change this to whatever extent you need. Format is W Longitude,S Latitude,E Longitude,N Latitude
     bounding_extent = args.bbox
@@ -245,7 +268,8 @@ def run():
 
     if defined_time_range:
         # if(data_since):
-        temporal_range = pa.get_temporal_range(start_date_time, end_date_time, datetime.utcnow().strftime("%Y-%m-%dT%H:%M:%SZ"))  # noqa E501
+        temporal_range = pa.get_temporal_range(start_date_time, end_date_time,
+                                               datetime.utcnow().strftime("%Y-%m-%dT%H:%M:%SZ"))  # noqa E501
 
     params = {
         'scroll': "true",
@@ -272,16 +296,26 @@ def run():
         }
 
         if args.verbose:
-            print("Temporal Range: " + temporal_range)
+            logging.info("Temporal Range: " + temporal_range)
 
     if args.verbose:
-        print("Provider: " + provider)
-        print("Updated Since: " + data_within_last_timestamp)
+        logging.info("Provider: " + provider)
+        logging.info("Updated Since: " + data_within_last_timestamp)
 
-    results = pa.get_search_results(args, params)
+    # If 401 is raised, refresh token and try one more time
+    try:
+        results = pa.get_search_results(args, params)
+    except HTTPError as e:
+        if e.code == 401:
+            token = pa.refresh_token(token, 'podaac-subscriber')
+            params['token'] = token
+            results = pa.get_search_results(args, params)
+        else:
+            raise e
 
     if args.verbose:
-        print(str(results['hits'])+" new granules found for "+short_name+" since "+data_within_last_timestamp)   # noqa E501
+        logging.info(str(results[
+                             'hits']) + " new granules found for " + short_name + " since " + data_within_last_timestamp)  # noqa E501
 
     if any([args.dy, args.dydoy, args.dymd]):
         file_start_times = pa.parse_start_times(results)
@@ -291,8 +325,12 @@ def run():
     timestamp = datetime.utcnow().strftime("%Y-%m-%dT%H:%M:%SZ")
 
     downloads_all = []
-    downloads_data = [[u['URL'] for u in r['umm']['RelatedUrls'] if u['Type'] == "GET DATA" and ('Subtype' not in u or u['Subtype'] != "OPENDAP DATA")] for r in results['items']]
-    downloads_metadata = [[u['URL'] for u in r['umm']['RelatedUrls'] if u['Type'] == "EXTENDED METADATA"] for r in results['items']]
+    
+    downloads_data = [[u['URL'] for u in r['umm']['RelatedUrls'] if
+                       u['Type'] == "GET DATA" and ('Subtype' not in u or u['Subtype'] != "OPENDAP DATA")] for r in
+                      results['items']]
+    downloads_metadata = [[u['URL'] for u in r['umm']['RelatedUrls'] if u['Type'] == "EXTENDED METADATA"] for r in
+                          results['items']]
     checksums = extract_checksums(results)
 
     for f in downloads_data:
@@ -303,7 +341,8 @@ def run():
     downloads = [item for sublist in downloads_all for item in sublist]
 
     if len(downloads) >= page_size:
-        print("Warning: only the most recent " + str(page_size) + " granules will be downloaded; try adjusting your search criteria (suggestion: reduce time period or spatial region of search) to ensure you retrieve all granules.")
+        logging.warning("Only the most recent " + str(
+            page_size) + " granules will be downloaded; try adjusting your search criteria (suggestion: reduce time period or spatial region of search) to ensure you retrieve all granules.")
 
     # filter list based on extension
     if not extensions:
@@ -319,9 +358,9 @@ def run():
     # https://github.com/podaac/data-subscriber/issues/33
     # Make this a non-verbose message
     # if args.verbose:
-    print("Found " + str(len(downloads)) + " total files to download")
+    logging.info("Found " + str(len(downloads)) + " total files to download")
     if args.verbose:
-        print("Downloading files with extensions: " + str(extensions))
+        logging.info("Downloading files with extensions: " + str(extensions))
 
     # NEED TO REFACTOR THIS, A LOT OF STUFF in here
     # Finish by downloading the files to the data directory in a loop.
@@ -342,18 +381,17 @@ def run():
 
             # decide if we should actually download this file (e.g. we may already have the latest version)
             if(exists(output_path) and not args.force and checksum_does_match(output_path, checksums)):
-                print(str(datetime.now()) + " SKIPPED: " + f)
+                logging.info(str(datetime.now()) + " SKIPPED: " + f)
                 skip_cnt += 1
                 continue
 
             urlretrieve(f, output_path)
             pa.process_file(process_cmd, output_path, args)
-            print(str(datetime.now()) + " SUCCESS: " + f)
+            logging.info(str(datetime.now()) + " SUCCESS: " + f)
             success_cnt = success_cnt + 1
-        except Exception as e:
-            print(str(datetime.now()) + " FAILURE: " + f)
+        except Exception:
+            logging.warning(str(datetime.now()) + " FAILURE: " + f, exc_info=True)
             failure_cnt = failure_cnt + 1
-            print(e)
 
     # If there were updates to the local time series during this run and no
     # exceptions were raised during the download loop, then overwrite the
@@ -364,13 +402,27 @@ def run():
             with open(data_path + "/.update__" + short_name, "w") as f:
                 f.write(timestamp)
 
-    print("\nDownloaded Files: " + str(success_cnt))
-    print("Failed Files:     " + str(failure_cnt))
-    print("Skipped Files:    " + str(skip_cnt) + "\n")
+    logging.info("Downloaded Files: " + str(success_cnt))
+    logging.info("Failed Files:     " + str(failure_cnt))
+    logging.info("Skipped Files:    " + str(skip_cnt))
     pa.delete_token(token_url, token)
-    print("\nEND\n\n")
+    logging.info("END\n\n")
     exit(0)
 
 
+def main():
+    log_level = os.environ.get('PODAAC_LOGLEVEL', 'INFO').upper()
+    logging.basicConfig(stream=sys.stdout,
+                        format='[%(asctime)s] {%(filename)s:%(lineno)d} %(levelname)s - %(message)s',
+                        level=log_level)
+    logging.debug("Log level set to " + log_level)
+
+    try:
+        run()
+    except Exception as e:
+        logging.exception("Uncaught exception occurred during execution.")
+        exit(hash(e))
+
+
 if __name__ == '__main__':
-    run()
+    main()