+++ CHANGELOG.md
@@ -4,10 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 
-## Unreleased
+## [Unreleased]
 ### Added
 - check if file exists before downloading a file. [17](https://github.com/podaac/data-subscriber/issues/17)
 ### Changed
+- Retry CMR queries on server error using random exponential backoff max 60 seconds and 10 retries
+- Refresh token if CMR returns 401 error
+- Converted print statements to log statements
 ### Deprecated
 ### Removed
 ### Fixed